Move ErrorDialog from `error` package to `dialog`
