+++ app/app/src/main/kotlin/com/hedvig/app/feature/crossselling/ui/detail/CrossSellDetailScreen.kt
@@ -44,7 +44,7 @@ import com.hedvig.android.core.designsystem.component.list.SectionTitle
 import com.hedvig.android.core.designsystem.preview.HedvigPreview
 import com.hedvig.android.core.designsystem.theme.HedvigTheme
 import com.hedvig.android.core.ui.appbar.FadingTopAppBar
-import com.hedvig.android.core.ui.error.ErrorDialog
+import com.hedvig.android.core.ui.dialog.ErrorDialog
 import com.hedvig.android.core.ui.preview.rememberPreviewImageLoader
 import com.hedvig.android.core.ui.scaffold.Scaffold
 import com.hedvig.app.feature.crossselling.ui.CrossSellData
