+++ app/core/ui/src/main/kotlin/com/hedvig/android/core/ui/dialog/ErrorDialog.kt
@@ -1,4 +1,4 @@
-package com.hedvig.android.core.ui.error
+package com.hedvig.android.core.ui.dialog
 
 import androidx.compose.material3.AlertDialog
 import androidx.compose.material3.MaterialTheme
