+++ app/feature/changeaddress/src/main/kotlin/com/hedvig/android/feature/changeaddress/destination/ChangeAddressEnterNewAddressDestination.kt
@@ -45,7 +45,7 @@ import com.hedvig.android.core.designsystem.preview.HedvigPreview
 import com.hedvig.android.core.designsystem.theme.HedvigTheme
 import com.hedvig.android.core.ui.ValidatedInput
 import com.hedvig.android.core.ui.clearFocusOnTap
-import com.hedvig.android.core.ui.error.ErrorDialog
+import com.hedvig.android.core.ui.dialog.ErrorDialog
 import com.hedvig.android.core.ui.scaffold.HedvigScaffold
 import com.hedvig.android.feature.changeaddress.ChangeAddressUiState
 import com.hedvig.android.feature.changeaddress.ChangeAddressViewModel
